Support optional callback for when watch begins
For testing purposes, it is especially useful to know when Watchful has
set up the watch. This commit adds support to optionally pass a function
to `watchful/watch` that will be called when the watch is ready.

In addition, this commit includes fixes for bugs relating to
non-existent paths to watch and paths to files to watch (rather than
directories).