Deduplicate events when using FSEvents
Unfortunately, the events that are captured using FSEvents contain
'duplicates' (in some cases this may simply be the parent directory of a
file that has been changed). This commit adds code to the fse backend
aimed at deduplicating events. Because it works by ignoring paths for
parent directories of a longer path, it will cause events to be thrown
away which were separately triggered for that longer path but which were
collated together with other events. Setting latency to zero does not
stop this collation from occurring.